This implements updated vertical tab flyout from #82
