+++ Fastedit/Controls/VerticalTabsFlyoutControl.xaml
@@ -1,0 +1,43 @@
+﻿<Flyout
+    x:Class="Fastedit.Controls.VerticalTabsFlyoutControl"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Fastedit.Controls"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:tab="using:Fastedit.Tab"
+    mc:Ignorable="d"
+    Opened="ShowAllTabsFlyout_Opened" ShowMode="TransientWithDismissOnPointerMoveAway" AllowFocusOnInteraction="True" LightDismissOverlayMode="Off" Placement="Bottom">
+    <Grid CharacterReceived="AllTabsFlyout_CharacterReceived" PreviewKeyDown="AllTabsFlyout_KeyDown" HorizontalAlignment="Stretch" VerticalAlignment="Top">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="auto"/>
+            <RowDefinition Height="40"/>
+        </Grid.RowDefinitions>
+        <ListView MaxHeight="400" x:Name="listView" VerticalAlignment="Stretch" UseSystemFocusVisuals="False" MaxWidth="250" SelectionChanged="AllTabsFlyout_SelectionChanged">
+            <ListView.ItemTemplate>
+                <DataTemplate x:DataType="local:TabFlyoutItem">
+                    <Grid Width="200" Height="40" VerticalAlignment="Center" ToolTipService.ToolTip="{x:Bind Tab.DatabaseItem.FilePath}">
+                        <Grid.ColumnDefinitions>
+                            <ColumnDefinition Width="auto"/>
+                            <ColumnDefinition Width="*"/>
+                            <ColumnDefinition Width="auto"/>
+                        </Grid.ColumnDefinitions>
+                        <FontIcon Margin="0,0,10,0" Grid.Column="0" Glyph="&#xE8A5;"/>
+                        <TextBlock Grid.Column="1" Text="{x:Bind Tab.DatabaseItem.FileName}" x:Phase="1" VerticalAlignment="Center"/>
+                        <Button Tag="{x:Bind Tab}" HorizontalAlignment="Right" Height="30" BorderBrush="Transparent" Background="Transparent" Grid.Column="1" Width="30" Margin="5,5,0,5" Padding="0" Click="AllTabsFlyout_CloseTab">
+                            <SymbolIcon Symbol="Cancel"/>
+                        </Button>
+                    </Grid>
+                </DataTemplate>
+            </ListView.ItemTemplate>
+        </ListView>
+        <Button Width="220" HorizontalAlignment="Stretch" Grid.Row="1" Click="NewFile_Click">
+            <Button.Content>
+                <StackPanel Orientation="Horizontal" HorizontalAlignment="Stretch">
+                    <SymbolIcon Grid.Column="0" Symbol="Add"/>
+                    <TextBlock Grid.Column="1" Margin="10,0,0,0" Text="New tab"/>
+                    <TextBlock Grid.Column="2" Margin="10,0,0,0" Foreground="{ThemeResource SystemColorGrayTextBrush}" Text="Ctrl + T"/>
+                </StackPanel>
+            </Button.Content>
+        </Button>
+    </Grid>
+</Flyout>
